Fetch Post
What is inside:
Index23home.html to show a table with car model and brand;
Add23home.html, for a form with inputs. When info is sent from there - it appears in index.html;
Notification added which alerts a message when information from form is sucessfully sent;
Navigation between sites.